+++ config/services.yaml
@@ -23,7 +23,10 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
-    App\Services\API\LOL\GeneralApi:
+    App\Services\API\LOL\DataDragon\GeneralApi:
+        public: true
+
+    App\Services\API\LOL\DataDragon\ChampionApi:
         public: true
     
     App\Services\API\LOL\MatchApi:
