+++ src/Services/API/LOL/DataDragon/ChampionApi.php
@@ -1,0 +1,72 @@
+<?php
+
+namespace App\Services\API\LOL\DataDragon;
+
+use App\Services\API\LOL\BaseApi;
+use App\Services\API\LOL\Model\DataDragon\Champion;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+
+class ChampionApi
+{
+    private const URL_CHAMPIONS = "http://ddragon.leagueoflegends.com/cdn/{version}/data/{lang}/champion.json";
+
+    private BaseApi $baseApi;
+    private DenormalizerInterface $denormalizer;
+
+    public function __construct(
+        BaseApi $baseApi,
+        DenormalizerInterface $denormalizer
+    ) {
+        $this->baseApi = $baseApi;
+        $this->denormalizer = $denormalizer;
+    }
+
+    /**
+     * @return array<array-key,Champion>|null
+     */
+    public function champions(string $version, string $lang): ?array
+    {
+        if (strlen($version) <= 0 || strlen($lang) <= 0) {
+            return null;
+        }
+
+        $url = $this->baseApi->constructUrl(
+            self::URL_CHAMPIONS,
+            [
+                "version" => $version,
+                "lang" => $lang
+            ]
+        );
+
+        /** @var array<array-key,array<array<string, int|string>>>|null $champions */
+        $champions = $this->baseApi->callApi(
+            $url,
+            Request::METHOD_GET,
+        );
+
+        if (is_null($champions)) {
+            return null;
+        }
+
+        /** @var array<array-key,Champion>|null $championsDenormalize */
+        $championsDenormalize = $this->denormalizeArray($champions['data'], Champion::class);
+        return $championsDenormalize;
+    }
+
+    /**
+     * @param array<array-key,array<string,int|string>> $datas
+     * @param string $type
+     * @return array<array-key,Champion>
+     */
+    private function denormalizeArray(array $datas, string $type)
+    {
+        if (is_array($datas)) {
+            $listDataDenormalize = [];
+            foreach ($datas as $key => $data) {
+                $listDataDenormalize[] = $this->denormalizer->denormalize($data, $type);
+            }
+            return $listDataDenormalize;
+        }
+    }
+}
